1. mainbanner 와 contentsPb.js 통합
2. 자세히 보기 클릭시 로그인페이지 안가는 문제 해결
